moving email sender logic to a new service + tests
