+++ app/services/emailSender.js
@@ -1,0 +1,18 @@
+const nodemailer = require('nodemailer'),
+  errors = require('../errors'),
+  config = require('./../../config');
+
+const transporter = nodemailer.createTransport({
+  host: config.common.mailer.host || 'smtp.mailtrap.io',
+  port: config.common.mailer.port || 2525,
+  auth: {
+    user: config.common.mailer.auth.user || 'b3153d829ec625',
+    pass: config.common.mailer.auth.pass || '37818ff837057d'
+  }
+});
+
+exports.sendMail = (mailOptions) => {
+  return transporter.sendMail(mailOptions).catch((err) => {
+    throw errors.defaultError(err.message);
+  });
+};
