+++ docs/index.html
@@ -1,1 +1,1 @@
-<html><head><link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" /><script src="https://code.jquery.com/jquery-1.12.2.min.js"></script><script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script><style> pre { width: 95%; } </style></head><body><div class="container"><div class="jumbotron"><h1>Documentation</h1></div><h1>Resources</h1><ul><li><a undefined href="games.html">Games</a></li><li><a undefined href="users.html">Users</a></li><li><a undefined href="admin.html">Admin</a></li></ul></div></body></html>
+<html><head><link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" /><script src="https://code.jquery.com/jquery-1.12.2.min.js"></script><script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script><style> pre { width: 95%; } </style></head><body><div class="container"><div class="jumbotron"><h1>Documentation</h1></div><h1>Resources</h1><ul><li><a undefined href="users.html">Users</a></li></ul></div></body></html>
