+++ docs/users.html
@@ -1,0 +1,22 @@
+<html><head><link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" /><script src="https://code.jquery.com/jquery-1.12.2.min.js"></script><script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script><style> pre { width: 95%; } </style></head><body><div class="container"><div class="jumbotron"><h1>Users</h1></div><div class="panel panel-default"><div class="panel-heading" role="tab" id="heading-users-1-matches-emailSummary-GET"><h4 class="panel-title"><a role="button" data-toggle="collapse" data-parent="#accordion" aria-expanded="true" aria-controls="collapse-users-1-matches-emailSummary-GET" href="#collapse-users-1-matches-emailSummary-GET">GET /users/1/matches/emailSummary</a></h4></div><div id="collapse-users-1-matches-emailSummary-GET" class="panel-collapse collapse" role="tabpanel" aria-labelledby="heading-users-1-matches-emailSummary-GET"><div class="panel-body"><h4>Request Data</h4><div class="container"><h5>Headers</h5><pre>{
+  "authorization": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdXRoZW50aWNhdGlvbkNvZGUiOiIkMmEkMTAkUG5tUFNzZVh5aHB6SDFxckJLcVJKLiIsIm1heGltdW1Vc2VmdWxEYXRlIjoiMjAxNy0wOC0yN1QxODozMDowMi40NDlaIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDE3LTA3LTMwVDE4OjMwOjAyLjQ1MVoiLCJpZCI6MSwicmVuZXdJZCI6IiQyYSQxMCRJRjRidTlEMGdWazBHZmlCOEF5YmZPIn0.n5pB9YnFgl6iPuJMxny8WnGkTP0f-UsIWGPAgzBcuuY"
+}</pre></div></div><div class="panel-body"><h4>Response Data</h4><div class="container"><h5>Status: 200</h5><h5>Headers</h5><pre>{
+  "authorization": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdXRoZW50aWNhdGlvbkNvZGUiOiIkMmEkMTAkUG5tUFNzZVh5aHB6SDFxckJLcVJKLiIsIm1heGltdW1Vc2VmdWxEYXRlIjoiMjAxNy0wOC0yN1QxODozMDowMi40NDlaIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDE3LTA3LTMwVDE4OjMwOjAyLjQ1MVoiLCJpZCI6MSwicmVuZXdJZCI6IiQyYSQxMCRJRjRidTlEMGdWazBHZmlCOEF5YmZPIn0.n5pB9YnFgl6iPuJMxny8WnGkTP0f-UsIWGPAgzBcuuY"
+}</pre><h5>Body</h5><pre>[
+  {
+    "id": 1,
+    "user_id": 1,
+    "game_id": 2,
+    "assertions": 6,
+    "createdAt": "2017-07-28T18:30:02.203Z",
+    "updatedAt": "2017-07-28T18:30:02.203Z"
+  },
+  {
+    "id": 3,
+    "user_id": 1,
+    "game_id": 1,
+    "assertions": 3,
+    "createdAt": "2017-07-28T18:30:02.204Z",
+    "updatedAt": "2017-07-28T18:30:02.204Z"
+  }
+]</pre></div></div></div></div></div></body></html>
